examples/google_adk/requirements.txt
google-adk>=1.27.2
pydantic>=2.0
